+++ examples/todos/lib/todos_page.dart
@@ -9,10 +9,10 @@ class TodosPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // Using SolidProvider here to provide the [TodosController] to descendants.
+    // Using Provider here to provide the [TodosController] to descendants.
     return Solid(
       providers: [
-        SolidProvider<TodosController>(
+        Provider<TodosController>(
           create: () => TodosController(initialTodos: Todo.sample),
           dispose: (controller) => controller.dispose(),
         ),