+++ packages/solidart_lint/example/lib/main.dart
@@ -18,12 +18,12 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return Solid(
       providers: [
-        // expect_lint: avoid_dynamic_solid_provider
-        SolidProvider(create: () => MyClass()),
-        // expect_lint: avoid_dynamic_solid_signal
-        SolidSignal(create: () => Signal(0), id: 'counter'),
-        // expect_lint: avoid_dynamic_solid_signal
-        SolidSignal(
+        // expect_lint: avoid_dynamic_provider
+        Provider(create: () => MyClass()),
+        // expect_lint: avoid_dynamic_provider
+        Provider(create: () => Signal(0), id: 'counter'),
+        // expect_lint: avoid_dynamic_provider
+        Provider(
           create: () => Computed(() => counter() * 2),
           id: 'double-counter',
         ),